requirements.txt
Django==4.0.5
django_filter==2.4.0
python-decouple==3.8
